Extract extension function for parsing XPath to Query
